Add MVP for BooksViewController
