+++ iosStudying/iosStudying/Flows/Books/BooksViewController.swift
@@ -9,6 +9,8 @@ import UIKit
 
 class BooksViewController: UIViewController {
 
+    var presenter: BooksPresenter?
+
     lazy var tableView: UITableView = {
         let tableView = UITableView()
         tableView.dataSource = self
@@ -17,8 +19,6 @@ class BooksViewController: UIViewController {
         return tableView
     }()
 
-    var books: BookResponse?
-    let service = BookService()
     var onFinishFlow: (() -> Void)?
     var onChapters: ((String, String) -> Void)?
 
@@ -31,43 +31,25 @@ class BooksViewController: UIViewController {
 
         self.title = R.string.localizible.booksViewControllerTitle()
         configureExitButton()
-        getData()
-    }
-
-    func getData() {
-        service.fetchBooks { result in
-            switch result {
-            case .success(let books):
-                self.books = books
-                self.tableView.reloadData()
-            case .failure(let error):
-                print(error)
-            }
-        }
+        presenter?.getData()
     }
 
     func configureExitButton() {
-        let barButton = UIBarButtonItem(title: R.string.localizible.logoutButtonTitle(),
-                                        style: .plain,
-                                        target: self,
-                                        action: #selector(isLogoutButtonPressed))
-        self.navigationItem.leftBarButtonItem = barButton
-    }
-
-    @objc private func isLogoutButtonPressed() {
-        UserDefaults.standard.set(false, forKey: PublicConstants.authKey)
-        onFinishFlow?()
+        self.navigationItem.leftBarButtonItem = UIBarButtonItem(title: R.string.localizible.logoutButtonTitle(),
+                                                                style: .plain,
+                                                                target: self,
+                                                                action: #selector(presenter?.finishFlow))
     }
 }
 
 extension BooksViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return books?.docs.count ?? 0
+        return presenter?.books?.docs.count ?? 0
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: UITableViewCell.description()),
-              let book = books else { return UITableViewCell() }
+              let book = presenter?.books else { return UITableViewCell() }
         cell.textLabel?.text = book.docs[indexPath.row].name
         return cell
     }
@@ -76,7 +58,7 @@ extension BooksViewController: UITableViewDataSource {
 extension BooksViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
-        if let book  = books?.docs[indexPath.row] {
+        if let book  = presenter?.books?.docs[indexPath.row] {
             onChapters?(book.identifier, book.name)
         }
     }
